requirements.txt
requests==2.32
python-dotenv